Localisation updates from http://translatewiki.net.
Codes that fall back (for example sr to sr-ec) are now filled, and code code they fall back to are removed. There remains a sorting issue (causing fiu-vro to be sorted at vro), but this is much clearer than it was.
Rev 112439 in MediaWiki updated the code map for this (https://www.mediawiki.org/wiki/Special:Code/MediaWiki/112439).

https://mediawiki.org/wiki/Special:Code/pywikipedia/9937