[IMPR] Modify 'unusedfiles' bot summary
The current summary for the unusedfiles.py is misleading. An orphan file
does not necesarily mean that the file will be deleted. Yet the summary
that this script uses ("images for elimination") implies that. Modern
wiki installations also do allow nowadays the uploading of several types
of multimedia files. Therefore proposing a more accurate "tagging file
as orphaned" summary for this script. This patch also mades a syntax
change, changing spaces to use tabs.

Change-Id: I23c4d3717c7507b6c007748bda156806bb854f4f